Add TokenLong64 class
Introduces new TokenLong64 class, which is an abstraction of all tokens
that can be represented as 64-bit integers. Change ShardingInfo to take
advantage of this new class.
